refactor: validateBookable 책임 변경
